refactor key vault - only keychain option available
