+++ agent-pkg/src/config/default.ts
@@ -1,15 +1,10 @@
-
-
 export const DEFAULT_CONFIG = {
     autoDriveNetwork: 'mainnet',
     taurusRpcUrl: 'https://auto-evm.taurus.autonomys.xyz/ws',
     packageRegistryAddress: '0x0B5cF4C198E8c75e8fE9B4D33F0B29881D13744b',
     autoSaveCredentials: true,
-    useKeychain: true,
   };
   
   // Service name for keytar (system keychain)
   export const KEYCHAIN_SERVICE = 'autoOS-cli';
   export const KEYCHAIN_ACCOUNT = 'masterPassword';
-
-  export const PASSWORD_CACHE_TIMEOUT = 10 * 60 * 1000;
